requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
mysqlclient==2.2.4
rest-framework-simplejwt==0.0.2
sqlparse==0.5.0
tzdata==2024.1
